+++ zcash_client_backend/CHANGELOG.md
@@ -12,8 +12,10 @@ and this library adheres to Rust's notion of
   - `BlockMetadata::orchard_tree_size`.
   - `TransparentInputSource`
   - `SaplingInputSource`
-  - `ScannedBlock::sapling_tree_size`.
-  - `ScannedBlock::orchard_tree_size`.
+  - `ScannedBlock::{
+      sapling_tree_size, orchard_tree_size, orchard_nullifier_map,
+      orchard_commitments, into_commitments
+    }`
   - `wallet::propose_standard_transfer_to_address`
   - `wallet::input_selection::Proposal::from_parts`
   - `wallet::input_selection::SaplingInputs`
@@ -164,6 +166,9 @@ and this library adheres to Rust's notion of
   removed without replacement as it was unused, and its functionality will be
   fully reproduced by `SaplingInputSource::select_spendable_sapling_notes` in a future
   change.
+- `zcash_client_backend::data_api::ScannedBlock::into_sapling_commitments` has been
+  replaced by `into_commitments` which returns both Sapling and Orchard note commitments
+  and associated note commitment retention information for the block.
 
 ## [0.10.0] - 2023-09-25